Got working graphql server and api routes running, need to connect to db and frontend
