+++ package.json
@@ -58,6 +58,8 @@
     "graphql-tag": "^2.10.1",
     "graphql-yoga": "^1.18.3",
     "history": "^4.9.0",
+    "neo4j-driver": "^1.7.6",
+    "neo4j-graphql-js": "^2.9.3",
     "parcelmon": "0.0.9",
     "react": "16.8.3",
     "react-apollo": "^3.1.3",
