--- server/index2.js
@@ -1,66 +1,0 @@
-const path = require('path')
-const {ApolloServer} = require('apollo-server-express')
-const express = require('express')
-const morgan = require('morgan')
-const compression = require('compression')
-const typeDefs = require('./schema')
-const Query = require('./resolvers/Query')
-const Mutation = require('./resolvers/Mutation')
-const Subscription = require('./resolvers/Subscription')
-const User = require('./resolvers/User')
-const PORT = process.env.PORT || 1234
-const app = express()
-module.exports = app
-
-const resolvers = {
-  Query,
-  Mutation,
-  Subscription,
-  User
-}
-
-const createApp = () => {
-  // logging middleware
-  app.use(morgan('dev'))
-
-  // body parsing middleware
-  app.use(express.json())
-  app.use(express.urlencoded({extended: true}))
-
-  // compression middleware
-  app.use(compression())
-
-  // any remaining requests with an extension (.js, .css, etc.) send 404
-  app.use((req, res, next) => {
-    if (path.extname(req.path).length) {
-      const err = new Error('Not found')
-      err.status = 404
-      next(err)
-    } else {
-      next()
-    }
-  })
-
-  server.applyMiddleware({app})
-
-  // error handling endware
-  app.use((err, req, res) => {
-    console.error(err)
-    console.error(err.stack)
-    res.status(err.status || 500).send(err.message || 'Internal server error.')
-  })
-}
-
-const startListening = () => {
-  // start listening (and create a 'server' object representing our server)
-  const server = app.listen(PORT, () =>
-    console.log(`Mixing it up on port ${PORT}`)
-  )
-}
-
-async function bootApp() {
-  await createApp()
-  await startListening()
-}
-
-bootApp()
